Moved directores and added Blade directive for specific roles
Advertisers only see CRUD operations, Viewers can only View stuff
